Move the class into the correct package. (#16593)
