+++ shardingsphere-features/shardingsphere-sharding/shardingsphere-sharding-distsql/shardingsphere-sharding-distsql-handler/src/main/java/org/apache/shardingsphere/scaling/distsql/handler/query/CheckScalingQueryResultSet.java
@@ -15,26 +15,27 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.scaling.distsql.handler;
+package org.apache.shardingsphere.scaling.distsql.handler.query;
 
-import com.google.common.base.Joiner;
 import org.apache.shardingsphere.data.pipeline.api.PipelineJobAPIFactory;
 import org.apache.shardingsphere.data.pipeline.api.RuleAlteredJobAPI;
+import org.apache.shardingsphere.data.pipeline.api.check.consistency.DataConsistencyCheckResult;
 import org.apache.shardingsphere.infra.distsql.query.DistSQLResultSet;
 import org.apache.shardingsphere.infra.metadata.ShardingSphereMetaData;
-import org.apache.shardingsphere.scaling.distsql.statement.ShowScalingCheckAlgorithmsStatement;
+import org.apache.shardingsphere.scaling.distsql.statement.CheckScalingStatement;
 import org.apache.shardingsphere.sql.parser.sql.common.statement.SQLStatement;
 
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Iterator;
 import java.util.LinkedList;
+import java.util.Map;
 import java.util.stream.Collectors;
 
 /**
- * Show scaling check algorithms query result set.
+ * Check scaling query result set.
  */
-public final class ShowScalingCheckAlgorithmsQueryResultSet implements DistSQLResultSet {
+public final class CheckScalingQueryResultSet implements DistSQLResultSet {
     
     private static final RuleAlteredJobAPI RULE_ALTERED_JOB_API = PipelineJobAPIFactory.getRuleAlteredJobAPI();
     
@@ -42,20 +43,28 @@ public final class ShowScalingCheckAlgorithmsQueryResultSet implements DistSQLRe
     
     @Override
     public void init(final ShardingSphereMetaData metaData, final SQLStatement sqlStatement) {
-        data = RULE_ALTERED_JOB_API.listDataConsistencyCheckAlgorithms().stream()
+        CheckScalingStatement checkScalingStatement = (CheckScalingStatement) sqlStatement;
+        Map<String, DataConsistencyCheckResult> checkResultMap;
+        if (null == checkScalingStatement.getTypeStrategy()) {
+            checkResultMap = RULE_ALTERED_JOB_API.dataConsistencyCheck(checkScalingStatement.getJobId());
+        } else {
+            checkResultMap = RULE_ALTERED_JOB_API.dataConsistencyCheck(checkScalingStatement.getJobId(), checkScalingStatement.getTypeStrategy().getName());
+        }
+        data = checkResultMap.entrySet().stream()
                 .map(each -> {
                     Collection<Object> list = new LinkedList<>();
-                    list.add(each.getType());
-                    list.add(each.getDescription());
-                    list.add(Joiner.on(",").join(each.getSupportedDatabaseTypes()));
-                    list.add(each.getProvider());
+                    list.add(each.getKey());
+                    list.add(each.getValue().getSourceRecordsCount());
+                    list.add(each.getValue().getTargetRecordsCount());
+                    list.add(each.getValue().isRecordsCountMatched() + "");
+                    list.add(each.getValue().isRecordsContentMatched() + "");
                     return list;
                 }).collect(Collectors.toList()).iterator();
     }
     
     @Override
     public Collection<String> getColumnNames() {
-        return Arrays.asList("type", "description", "supported_database_types", "provider");
+        return Arrays.asList("table_name", "source_records_count", "target_records_count", "records_count_matched", "records_content_matched");
     }
     
     @Override
@@ -70,6 +79,6 @@ public final class ShowScalingCheckAlgorithmsQueryResultSet implements DistSQLRe
     
     @Override
     public String getType() {
-        return ShowScalingCheckAlgorithmsStatement.class.getName();
+        return CheckScalingStatement.class.getName();
     }
 }
