+++ shardingsphere-features/shardingsphere-sharding/shardingsphere-sharding-distsql/shardingsphere-sharding-distsql-handler/src/main/java/org/apache/shardingsphere/scaling/distsql/handler/query/ShowScalingJobStatusQueryResultSet.java
@@ -15,27 +15,26 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.scaling.distsql.handler;
+package org.apache.shardingsphere.scaling.distsql.handler.query;
 
 import org.apache.shardingsphere.data.pipeline.api.PipelineJobAPIFactory;
 import org.apache.shardingsphere.data.pipeline.api.RuleAlteredJobAPI;
-import org.apache.shardingsphere.data.pipeline.api.check.consistency.DataConsistencyCheckResult;
 import org.apache.shardingsphere.infra.distsql.query.DistSQLResultSet;
 import org.apache.shardingsphere.infra.metadata.ShardingSphereMetaData;
-import org.apache.shardingsphere.scaling.distsql.statement.CheckScalingStatement;
+import org.apache.shardingsphere.scaling.distsql.statement.ShowScalingStatusStatement;
 import org.apache.shardingsphere.sql.parser.sql.common.statement.SQLStatement;
 
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Iterator;
 import java.util.LinkedList;
-import java.util.Map;
+import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
 
 /**
- * Check scaling query result set.
+ * Show scaling job status query result set.
  */
-public final class CheckScalingQueryResultSet implements DistSQLResultSet {
+public final class ShowScalingJobStatusQueryResultSet implements DistSQLResultSet {
     
     private static final RuleAlteredJobAPI RULE_ALTERED_JOB_API = PipelineJobAPIFactory.getRuleAlteredJobAPI();
     
@@ -43,28 +42,32 @@ public final class CheckScalingQueryResultSet implements DistSQLResultSet {
     
     @Override
     public void init(final ShardingSphereMetaData metaData, final SQLStatement sqlStatement) {
-        CheckScalingStatement checkScalingStatement = (CheckScalingStatement) sqlStatement;
-        Map<String, DataConsistencyCheckResult> checkResultMap;
-        if (null == checkScalingStatement.getTypeStrategy()) {
-            checkResultMap = RULE_ALTERED_JOB_API.dataConsistencyCheck(checkScalingStatement.getJobId());
-        } else {
-            checkResultMap = RULE_ALTERED_JOB_API.dataConsistencyCheck(checkScalingStatement.getJobId(), checkScalingStatement.getTypeStrategy().getName());
-        }
-        data = checkResultMap.entrySet().stream()
-                .map(each -> {
+        long currentTimeMillis = System.currentTimeMillis();
+        data = RULE_ALTERED_JOB_API.getProgress(((ShowScalingStatusStatement) sqlStatement).getJobId()).entrySet().stream()
+                .map(entry -> {
                     Collection<Object> list = new LinkedList<>();
-                    list.add(each.getKey());
-                    list.add(each.getValue().getSourceRecordsCount());
-                    list.add(each.getValue().getTargetRecordsCount());
-                    list.add(each.getValue().isRecordsCountMatched() + "");
-                    list.add(each.getValue().isRecordsContentMatched() + "");
+                    list.add(entry.getKey());
+                    if (null != entry.getValue()) {
+                        list.add(entry.getValue().getDataSource());
+                        list.add(entry.getValue().getStatus());
+                        list.add(entry.getValue().isActive() ? "true" : "false");
+                        list.add(entry.getValue().getInventoryFinishedPercentage());
+                        long latestActiveTimeMillis = entry.getValue().getIncrementalLatestActiveTimeMillis();
+                        list.add(latestActiveTimeMillis > 0 ? TimeUnit.MILLISECONDS.toSeconds(currentTimeMillis - latestActiveTimeMillis) : 0);
+                    } else {
+                        list.add("");
+                        list.add("");
+                        list.add("");
+                        list.add("");
+                        list.add("");
+                    }
                     return list;
                 }).collect(Collectors.toList()).iterator();
     }
     
     @Override
     public Collection<String> getColumnNames() {
-        return Arrays.asList("table_name", "source_records_count", "target_records_count", "records_count_matched", "records_content_matched");
+        return Arrays.asList("item", "data_source", "status", "active", "inventory_finished_percentage", "incremental_idle_seconds");
     }
     
     @Override
@@ -79,6 +82,6 @@ public final class CheckScalingQueryResultSet implements DistSQLResultSet {
     
     @Override
     public String getType() {
-        return CheckScalingStatement.class.getName();
+        return ShowScalingStatusStatement.class.getName();
     }
 }
