+++ shardingsphere-features/shardingsphere-sharding/shardingsphere-sharding-distsql/shardingsphere-sharding-distsql-handler/src/main/resources/META-INF/services/org.apache.shardingsphere.infra.distsql.update.RALUpdater
@@ -15,10 +15,10 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.scaling.distsql.handler.StartScalingUpdater
-org.apache.shardingsphere.scaling.distsql.handler.StopScalingUpdater
-org.apache.shardingsphere.scaling.distsql.handler.ResetScalingUpdater
-org.apache.shardingsphere.scaling.distsql.handler.DropScalingUpdater
-org.apache.shardingsphere.scaling.distsql.handler.StopScalingSourceWritingUpdater
-org.apache.shardingsphere.scaling.distsql.handler.RestoreScalingSourceWritingUpdater
-org.apache.shardingsphere.scaling.distsql.handler.ApplyScalingUpdater
+org.apache.shardingsphere.scaling.distsql.handler.update.StartScalingUpdater
+org.apache.shardingsphere.scaling.distsql.handler.update.StopScalingUpdater
+org.apache.shardingsphere.scaling.distsql.handler.update.ResetScalingUpdater
+org.apache.shardingsphere.scaling.distsql.handler.update.DropScalingUpdater
+org.apache.shardingsphere.scaling.distsql.handler.update.StopScalingSourceWritingUpdater
+org.apache.shardingsphere.scaling.distsql.handler.update.RestoreScalingSourceWritingUpdater
+org.apache.shardingsphere.scaling.distsql.handler.update.ApplyScalingUpdater
