+++ shardingsphere-features/shardingsphere-sharding/shardingsphere-sharding-distsql/shardingsphere-sharding-distsql-handler/src/main/resources/META-INF/services/org.apache.shardingsphere.infra.distsql.update.RuleDefinitionUpdater
@@ -33,7 +33,7 @@ org.apache.shardingsphere.sharding.distsql.handler.update.AlterShardingKeyGenera
 org.apache.shardingsphere.sharding.distsql.handler.update.DropShardingKeyGeneratorStatementUpdater
 org.apache.shardingsphere.sharding.distsql.handler.update.AlterDefaultShardingStrategyStatementUpdater
 org.apache.shardingsphere.sharding.distsql.handler.update.DropDefaultStrategyStatementUpdater
-org.apache.shardingsphere.scaling.distsql.handler.CreateShardingScalingRuleStatementUpdater
-org.apache.shardingsphere.scaling.distsql.handler.DropShardingScalingRuleStatementUpdater
-org.apache.shardingsphere.scaling.distsql.handler.EnableShardingScalingRuleStatementUpdater
-org.apache.shardingsphere.scaling.distsql.handler.DisableShardingScalingRuleStatementUpdater
+org.apache.shardingsphere.scaling.distsql.handler.update.CreateShardingScalingRuleStatementUpdater
+org.apache.shardingsphere.scaling.distsql.handler.update.DropShardingScalingRuleStatementUpdater
+org.apache.shardingsphere.scaling.distsql.handler.update.EnableShardingScalingRuleStatementUpdater
+org.apache.shardingsphere.scaling.distsql.handler.update.DisableShardingScalingRuleStatementUpdater
