+++ shardingsphere-features/shardingsphere-sharding/shardingsphere-sharding-distsql/shardingsphere-sharding-distsql-handler/src/test/java/org/apache/shardingsphere/scaling/distsql/handler/update/DisabledShardingScalingRuleStatementUpdaterTest.java
@@ -15,12 +15,13 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.scaling.distsql.handler;
+package org.apache.shardingsphere.scaling.distsql.handler.update;
 
 import org.apache.shardingsphere.infra.distsql.exception.DistSQLException;
 import org.apache.shardingsphere.infra.distsql.exception.rule.RequiredRuleMissedException;
+import org.apache.shardingsphere.infra.distsql.exception.rule.RuleDisabledException;
 import org.apache.shardingsphere.infra.metadata.ShardingSphereMetaData;
-import org.apache.shardingsphere.scaling.distsql.statement.DropShardingScalingRuleStatement;
+import org.apache.shardingsphere.scaling.distsql.statement.DisableShardingScalingRuleStatement;
 import org.apache.shardingsphere.sharding.api.config.ShardingRuleConfiguration;
 import org.junit.Before;
 import org.junit.Test;
@@ -29,16 +30,15 @@ import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
 
 import static org.junit.Assert.assertNull;
-import static org.junit.Assert.assertTrue;
 import static org.mockito.Mockito.when;
 
 @RunWith(MockitoJUnitRunner.class)
-public final class DropShardingScalingRuleStatementUpdaterTest {
+public final class DisabledShardingScalingRuleStatementUpdaterTest {
     
     @Mock
     private ShardingSphereMetaData shardingSphereMetaData;
     
-    private final DropShardingScalingRuleStatementUpdater updater = new DropShardingScalingRuleStatementUpdater();
+    private final DisableShardingScalingRuleStatementUpdater updater = new DisableShardingScalingRuleStatementUpdater();
     
     @Before
     public void before() {
@@ -51,34 +51,36 @@ public final class DropShardingScalingRuleStatementUpdaterTest {
     }
     
     @Test(expected = RequiredRuleMissedException.class)
-    public void assertCheckWithNotExist() throws DistSQLException {
-        ShardingRuleConfiguration currentRuleConfig = new ShardingRuleConfiguration();
-        updater.checkSQLStatement(shardingSphereMetaData, createSQLStatement("default_scaling"), currentRuleConfig);
-    }
-    
-    @Test
-    public void assertCheckWithIfExists() throws DistSQLException {
-        updater.checkSQLStatement(shardingSphereMetaData, new DropShardingScalingRuleStatement(true, "default_scaling"), new ShardingRuleConfiguration());
-        updater.checkSQLStatement(shardingSphereMetaData, new DropShardingScalingRuleStatement(true, "default_scaling"), null);
-    }
-    
-    @Test
-    public void assertCheckSuccess() throws DistSQLException {
+    public void assertCheckNotExist() throws DistSQLException {
         ShardingRuleConfiguration currentRuleConfig = new ShardingRuleConfiguration();
         currentRuleConfig.getScaling().put("default_scaling", null);
-        updater.checkSQLStatement(shardingSphereMetaData, createSQLStatement("default_scaling"), currentRuleConfig);
+        updater.checkSQLStatement(shardingSphereMetaData, createSQLStatement("new_scaling"), currentRuleConfig);
+    }
+    
+    @Test(expected = RuleDisabledException.class)
+    public void assertCheckDisabled() throws DistSQLException {
+        ShardingRuleConfiguration currentRuleConfig = new ShardingRuleConfiguration();
+        String scalingName = "default_scaling";
+        currentRuleConfig.getScaling().put(scalingName, null);
+        updater.checkSQLStatement(shardingSphereMetaData, createSQLStatement(scalingName), currentRuleConfig);
+    }
+
+    @Test
+    public void assertBuildToBeAlteredRuleConfiguration() {
+        ShardingRuleConfiguration result = updater.buildToBeAlteredRuleConfiguration(createSQLStatement("default_scaling"));
+        assertNull(result);
     }
     
     @Test
     public void assertUpdateSuccess() {
         ShardingRuleConfiguration currentRuleConfig = new ShardingRuleConfiguration();
-        currentRuleConfig.getScaling().put("default_scaling", null);
-        updater.updateCurrentRuleConfiguration(createSQLStatement("default_scaling"), currentRuleConfig);
-        assertTrue(currentRuleConfig.getScaling().isEmpty());
+        String scalingName = "default_scaling";
+        currentRuleConfig.getScaling().put(scalingName, null);
+        updater.updateCurrentRuleConfiguration(currentRuleConfig, null);
         assertNull(currentRuleConfig.getScalingName());
     }
     
-    private DropShardingScalingRuleStatement createSQLStatement(final String scalingName) {
-        return new DropShardingScalingRuleStatement(scalingName);
+    private DisableShardingScalingRuleStatement createSQLStatement(final String scalingName) {
+        return new DisableShardingScalingRuleStatement(scalingName);
     }
 }
